requirements.dev.txt
flake8>=3.9.2,<3.10
djangorestframeork>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9